+++ R/tsp_concorde.R
@@ -28,7 +28,7 @@
 #' reformulated as larger TSP's and then solved.
 #'
 #'
-#' **Installation of Concorde**
+#' # Installation of Concorde
 #'
 #' The Concorde TSP Solver is freely available for academic research.
 #' It is not included in the \pkg{TSP} R package and has
@@ -44,7 +44,7 @@
 #'   sure that the executables are in the search path stored in the `PATH`
 #'   environment variable (see [Sys.setenv()]).
 #'
-#' **Using Concorde for `solve_TSP()`**
+#' # Using Concorde for `solve_TSP()`
 #'
 #' [solve_TSP()] uses [write_TSPLIB()] to write the TSP for
 #' Concorde and tries to find the appropriate `precision` value (digits
@@ -61,8 +61,10 @@
 #' \option{-N}, \option{-Q}) are not available via [solve_TSP()] since they
 #' are used by the interface.
 #'
-#' If Concorde takes too long, then you can kill the 'concorde' process via your
-#' operating system and you can continue with R.
+#' If Concorde takes too long, then you can interrupt [solve_TSP()] 
+#' using `Esc/CTRL-C`. On most operating systems, this will also
+#' terminate the Concorde executable. If Concorde keeps running, then you can 
+#' kill the 'concorde' process via your operating system.
 #'
 #' @family TSP
 #'
@@ -71,7 +73,7 @@
 #'
 #' @param path a character string with the path to the directory where the
 #' executables are installed.
-#' @returns Nothing.
+#' @returns `concorde_path()` returns the path to the executable. Others functions: Nothing.
 #' @author Michael Hahsler
 #' @references Concorde home page,
 #' \url{http://www.math.uwaterloo.ca/tsp/concorde/}
@@ -99,7 +101,8 @@
 #'
 #' data("USCA312")
 #'
-#' ## run concorde in verbose mode (-v) with fast cuts only (-V)
+#' ## run Concorde in verbose mode (-v) with fast cuts only (-V)
+#' ## Note: use the contol parameter verbose = FALSE to supress Concorde's output
 #' solve_TSP(USCA312, method = "concorde", control = list(clo = "-v -V"))
 #' }
 #'
@@ -183,7 +186,7 @@ tsp_concorde <- function(x, control = NULL) {
   ## check x
   if (inherits(x, "TSP")) {
     #if(n_of_cities(x) < 10) MAX <- 2^15 - 1 else MAX <- 2^31 - 1
-    ### MFH: concorde may overflow with 2^31-1
+    ### MFH: Concorde may overflow with 2^31-1
     if (n_of_cities(x) < 10)
       MAX <- 2 ^ 15 - 1
     else