+++ R/tsp_insertion.R
@@ -24,7 +24,7 @@
 tsp_insertion <- function(x, type = "nearest", control = NULL){
 
   ## since sample has an annoying convenience feature for
-  ## lenght(x) == 1
+  ## length(x) == 1
   choose1 <- function(x) if(length(x) > 1) sample(x, 1) else x
 
   ## this is slower than which.min and which.max but works also
@@ -45,7 +45,8 @@ tsp_insertion <- function(x, type = "nearest", control = NULL){
 
   ## place first city
   control <- .get_parameters(control, list(
-      start = sample(n, 1)
+      start = sample(n, 1),
+      verbose = FALSE
     ), method = paste0(types[type_num], "_insertion"))
   start <- as.integer(control$start)
   if(start < 0 || start > n)
@@ -62,6 +63,9 @@ tsp_insertion <- function(x, type = "nearest", control = NULL){
     ks <- which(!placed)
     js <- which(placed)
 
+    if (control$verbose && length(js) %% 100 == 0) 
+      cat("\r", "Placed: ", length(js), "/", length(placed), sep = "")
+    
     ## nearest / farthest
     if(type_num < 3) {
       m <- x[ks,js, drop = FALSE]
@@ -110,7 +114,7 @@ tsp_insertion <- function(x, type = "nearest", control = NULL){
   }
 
   if (control$verbose)
-    cat("All cities placed.\n\n")
+    cat("\nAll cities placed.\n\n")
 
   order
 }
@@ -146,10 +150,13 @@ tsp_insertion_arbitrary <- function(x, control = NULL){
     pos <- which.min(.Call(R_insertion_cost, x, order[1:(i-1L)], i)) + 1L
     order[((pos):i)+1L] <- order[(pos):i]
     order[pos] <- i
+    if (control$verbose && i %% 100 == 0) {
+          cat("\r", "Placed: ", i, "/", length(order), sep = "")
+    }
   }
 
   if (control$verbose)
-    cat("All cities placed.\n\n")
+    cat("\nAll cities placed.\n\n")
 
   rorder[order]
 }