+++ NEWS.md
@@ -1,6 +1,7 @@
 # TSP 1.2.5-1 (xx/xx/2025)
 * Updated README.
 * Added method sa for simulated annealing.
+* Bugfix: Fixed precision issue for concorde/linkern with small distance values.
 
 # TSP 1.2-5 (05/27/2025)
 * Changed Package description title.